requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
python-multipart==0.0.12

pypdf==5.1.0
pdfminer.six==20240706

scikit-learn==1.5.2
numpy==2.1.2
scipy==1.13.1

openai==1.51.2
python-dotenv==1.0.1

pytest==8.3.3
